requirements.txt
requests>=2.22.0
python-telegram-bot>=12.1.1
python-dateutil>=2.8.0
pytz>=2019.2
PyYAML==5.1.2
